requirements.txt
openai>=1.40
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
langchain-text-splitters>=0.2
langchain-pinecone>=0.1
chromadb>=0.5
pinecone-client>=5.0.0
fastapi>=0.114.0
uvicorn[standard]>=0.30.0
streamlit>=1.36.0
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
tiktoken>=0.7.0
wandb>=0.17.0
pandas>=2.2.0
numpy>=2.0.0
tenacity>=8.2.3
httpx>=0.27.0
rich>=13.7.1
langsmith>=0.1.98
docx2txt>=0.8
pypdf>=4.2.0
